requirements.txt
pydantic-ai-slim[openai,vertexai]==0.0.43
fastapi[standard]==0.112.2
python-dotenv==1.0.1